[CL-2650] Copy ProjectCopyService into core and call from LocalProjectCopyService
[CL-2650] LocalProjectCopyService inherits from ProjectCopyService

[CL-2650] Remove original ProjectCopyService file from admin_api/services/ & updaterefs in specs
