+++ back/engines/commercial/admin_api/spec/services/project_copy_service_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe AdminApi::ProjectCopyService do
+describe ProjectCopyService do
   let(:service) { described_class.new }
 
   describe 'project copy', slow_test: true do
